ity to draw, add and remove shapes in GUI
This commit adds significant features to the FakePowerPoint application. Specifically, it implements the capability for the user interface to draw, add, and remove shapes.

The draw functionality is added with DrawRectangle and DrawLine methods in Form1 class and corresponding methods in IShape interface. Random shape coordinates now consider the GUI dimensions.

The add and remove shape operations are connected to the GUI with button click event handlers and integrated with data grid view for user interaction.

A new private variable of Model class has also been introduced in Form1 to keep track of the state. Every new shape is instantly drawn and data grid is updated accordingly.

A Random class instance is adjusted to be static to provide better randomization.

The commit adjusts color and size attributes of Line and Rectangle classes. Firstly, it modifies color attribute initialization. Secondly, ConvertToRectangle method is modified to consider coordinate differences for width and height as original version did not create the rectangle correctly.

Lastly, a ShapeFactory is implemented to create shape object depending on provided type string and GetCoordinates method is added in IShape interface which returns a string of the coordinates for the particular shape.

These all changes enhance the FakePowerPoint application by rendering it interactive and more user-friendly.