"Changed random number range for rectangle creation
Modified the range of random numbers generated when creating the bottom-right coordinates of Rectangle objects in FakePowerPoint program. Now, the bottom-right coordinate is guaranteed to be greater or equal to the top-left coordinate. This ensures that no Rectangle will have negative dimensions, which was causing errors in the rendering process."