Add "Info" menu to FakePowerPoint interface
This commit introduces an "Info" menu item on the main interface of our FakePowerPoint application. It also includes an "" submenu for users to get quick information about the app. Additionally, minor adjustments were made to the positioning and measurements of some elements for a more user-friendly interface. The changes made will improve usability and user experience.