Update cursor behavior in presentation model
The cursor's default state has been changed in the General.cs file of the PresentationModels directory. Previously, the cursor only returned to its default state if it was inside the PaintGroup. Now, the cursor will always be in its default state. The MouseMovingOnPanel() function is still conducted only if the cursor is inside the PaintGroup. This change was made to make the cursor behaviour more consistent and predictable for users.