"Add Normal Mode to Shape Selector and update data binding"
This commit introduces a new "normal" mode to the shape selector of our "Fake PowerPoint" application. It's addition is seen in the relevant button being added in the Form1.Designer.cs file.

The Normal Mode allows users to interact with the shapes already drawn on the canvas, without the chance of accidentally creating new ones, giving it an independent status from the rest of the shapes. Hence, it has been treated as a type of ShapeType.

Significant code changes include the addition of an event handler for the Normal Mode button in the Form1.cs and the handling logic for the selection update when Normal Mode is selected in the DataBindingThings.cs. The changes were made keeping scalability in mind, adding only minimal code that would integrate well with the existing structure.

The data binding logic has been updated to accommodate the 'undefined' mode within the general shape selection logic rather than treating it as a special case.

The selected shape for drawing is reset to 'undefined' after each shape draw in the Paint.cs file which means after each shape drawn, the application goes back to the normal mode, increasing the usability.

Lastly, the app initialization time is expected to decrease considering the disabling of double buffered, removing some potential for sluggish behaviour.