Add feature to bind slide background
A feature has been added to bind the slide background to a button in the PowerPoint Presentation. The changes are in the PresentationModels and the Form1 files. The new method 'BindSlideBackground' binds the slide background to a button. Method 'SetPaintGroup' now also includes code to draw the group box to a bitmap. Additional invalidation is now performed in the 'Draw' method to ensure the UI is updated correctly. 'button1' text is also updated for a better presentation of the button itself.