Refactor shapes to use Point and add Handle class
Refactored the Shapes classes Eclipse, Line, and Rectangle to use System.Drawing.Point instead of Tuple for coordinates. This simplifies code and increases readability.

A Handle class has also been introduced and added to the each shape. This will allow for easier manipulation of shapes in the UI in the future.

Extension methods for Draw and DrawHandle have been updated accordingly across all shapes to accommodate this change.