Improve rectangle shape handling and drawing
Modified the Rectangle and Handle classes in the Shapes module to improve handling and drawing of rectangles.
The change makes _shapeType attribute of Rectangle read-only for maintaining rectangle identity throughout its life.
Also, the Handles for the rectangle have been modified to include eight handles from the previously used two handles.
Additionally, the Draw function now renders the Handles when a rectangle is selected.
The handle drawing is now addressed inside the Handle class. This is a refactoring change to make drawing more versatile and dynamic, and closer to actual PowerPoint like behavior.