"Change private field '_coordinates' to public property in all shape classes"
This commit changes the '_coordinates' private field in all the shape classes (Eclipse, Line, Rectangle) to a public property 'Coordinates'. This change was made to standardize access to coordinates and provide the options to get and set them directly which were not allowed in the previous version. This change facilitates easier interaction with these coordinates, especially when