Refactor ShapeFactory to use factory method pattern
ShapeFactory was refactored to use factory method design pattern which improves the readability and the maintainability of the code. This also uses separation of concerns principle giving more structurally sound code.
Added interfaces for factory classes and injected them, following the Dependency Injection principle. This allows for easy future changes as the creation of new shape types only requires a new shape factory class.

Within the Model class, I also added functionality to use the new factories to create shapes. This allows us to encapsulate the logic of creating shapes in a separate class - we can easily extend our code with new shapes just by creating new Factory classes.

Additionally, code in the `Paint` class was refactored to use the new method of creating shapes provided by the Model class.

Coordinates have been changed from Tuple to list giving continuous collection. Now shapes are created based on a List of integers which makes it easier for the creation process and makes it flexible when more dimensions are needed.

Random coordinates are also incorporated for instantiating the model for generating the random shapes to be drawn on the canvas.

Finally, IShape interface has been updated for more clarity and encapsulation.