Replace "Undefined" with "Selection" and add new method descriptions
Updated the enum ShapeType by renaming the 'Undefined' value to 'Selection' to provide better clarity. Comments have also been added to describe the methods in the "Line" class. This will make the handling of shapes easier and the code more comprehensible. The logic to generate a new line in "LineFactory" has also been slightly modified for clarity.