"Replace Graphics with IGraphics for testability"
In this commit, every usage of System.Drawing.Graphics in the shapes classes was replaced with a new interface, IGraphics. This change was made to improve the code testability. Now we can implement a 'fake graphics' for unit test and separate the actual drawing from the application logic. Two IGraphics implementations were added: TheOnlyTrueGraphics, which encapsulates real drawing, and FakeGraphics, which only stores the calls and can be used in unit tests. The replacement affects shapes, handles, and the Paint class.