"Refactor project structure, remove redundant FakeGraphics class, add and update tests
The FakeGraphics class in the FakeUnitTest directory was removed as it contained not implemented methods. The already existing class in the Model folder is now used in all places.

Tests for the 'Rectangle' and 'ModelWithRects' classes were added into their respective Test files. This is to enhance our testing coverage and maintain code quality.

The 'Paint' class was updated to utilize the 'TheOnlyTrueGraphics' object, replacing direct use of Graphics to stick within established design patterns.

A setter was added to the FakeGraphics.cs file in Model folder to improve control over the Graphics object and facilitate testing.

Overall these changes enhance the cohesiveness and maintainability of the code while improving the testing coverage."