Update code formatting, add .gitignore
The commit includes a change in the formatting of multiple files in FakeUnitTest project to align with the standard code formatting rules. It also adds a .gitignore file to prevent unnecessary file tracking. The ignored files include asset files, binaries, object files and cache created by StyleCop.

These changes were made in order to conform to code style guidelines and clean up version control tracking, making it easier to read and understand the code and changes.