Refactor Slide class and improve graphics rendering
In this commit, necessary changes have been made in the Slide class to improve the rendering of shapes graphics. Now, each Slide instance keeps track of its own Size and Graphics object. These are now initialized when the Slide is created and updated when its size changes.

As a result, in the Draw() method, the Graphics object doesn't need to be created and disposed every time, but rather it is reused. These changes should improve the efficiency of drawing shapes and have a positive impact on application's performance.

The change triggers Repaint() and subsequently updates bound properties when AddShape() is called in PresentationModel.

Correctly, draws an Ellipse in the Ellipse class instead of a Rectangle.

The AddShape method is also changed in both PresentationModel and Model class, so it now uses the ShapeType enum directly rather than parsing a string. This makes the code a bit cleaner and less error-prone.

P.S: Avoid adding shapes with ShapeType.Undefined without checking validity.