Add new FakePowerPoint Presentation Form1.resx
Created a .resx file for Form1 in the FakePowerPoint project to define a set of resources to be used in the application. This includes UI elements and string definitions. This is done to enhance the localization support of the application, making it easier to translate the application into different languages.