Add undo/redo status tracking
An `ObservableList` named `_dos` was added to `PresentationModel` in order to track the status of undoable/redoable operations. It was initialized and updated in the appropriate methods to reflect these states. This addresses the need for intelligent enabling/disabling of 'Undo' and 'Redo' buttons in the user interface. Accordingly, the FormDesigner has been updated to initially disable these buttons and an EventHandler to enable/disable in line with `_dos` changes was added. These changes provide intuitive and efficient user interaction and prevent potential errors that could occur from invoking 'Undo'/'Redo' when there's nothing to be undone/redone.